+++ vue-project/src/services/api.ts
@@ -130,11 +130,8 @@ export function startSocket() {
         });
 
         socket.onAny((eventName, ...args) => {
-            console.log(args)
-            if (eventName == 'update_game') {
-                console.log(args[0]);
-                currentGame.value = args[0];
-            }
+            console.log(args[0]);
+            currentGame.value = args[0];
         });
     }
 }
